Update to Quartz 3.19

+++ src/Weasel.Quartz.Tests/quartz_postgressql.sql
@@ -1,30 +1,36 @@
-﻿set client_min_messages = WARNING;
-DROP TABLE IF EXISTS qrtz_fired_triggers;
-DROP TABLE IF EXISTS qrtz_paused_trigger_grps;
-DROP TABLE IF EXISTS qrtz_scheduler_state;
-DROP TABLE IF EXISTS qrtz_locks;
-DROP TABLE IF EXISTS qrtz_simprop_triggers;
-DROP TABLE IF EXISTS qrtz_simple_triggers;
-DROP TABLE IF EXISTS qrtz_cron_triggers;
-DROP TABLE IF EXISTS qrtz_blob_triggers;
-DROP TABLE IF EXISTS qrtz_triggers;
-DROP TABLE IF EXISTS qrtz_job_details;
-DROP TABLE IF EXISTS qrtz_calendars;
-set client_min_messages = NOTICE;
+﻿DO $$
+    DECLARE DropDb INT := 1; -- Set this to 0 to skip DROP statements, 1 to include them
+    BEGIN
+        IF DropDb = 1 THEN
+            SET client_min_messages = WARNING;
+            DROP TABLE IF EXISTS qrtz_fired_triggers;
+            DROP TABLE IF EXISTS qrtz_paused_trigger_grps;
+            DROP TABLE IF EXISTS qrtz_scheduler_state;
+            DROP TABLE IF EXISTS qrtz_locks;
+            DROP TABLE IF EXISTS qrtz_simprop_triggers;
+            DROP TABLE IF EXISTS qrtz_simple_triggers;
+            DROP TABLE IF EXISTS qrtz_cron_triggers;
+            DROP TABLE IF EXISTS qrtz_blob_triggers;
+            DROP TABLE IF EXISTS qrtz_triggers;
+            DROP TABLE IF EXISTS qrtz_job_details;
+            DROP TABLE IF EXISTS qrtz_calendars;
+            SET client_min_messages = NOTICE;
+        END IF;
+    END $$;
 
 CREATE TABLE qrtz_job_details
 (
     sched_name TEXT NOT NULL,
-    job_name  TEXT NOT NULL,
+    job_name TEXT NOT NULL,
     job_group TEXT NOT NULL,
     description TEXT NULL,
-    job_class_name   TEXT NOT NULL,
+    job_class_name TEXT NOT NULL,
     is_durable BOOL NOT NULL,
     is_nonconcurrent BOOL NOT NULL,
     is_update_data BOOL NOT NULL,
     requests_recovery BOOL NOT NULL,
     job_data BYTEA NULL,
-    PRIMARY KEY (sched_name,job_name,job_group)
+    PRIMARY KEY (sched_name, job_name, job_group)
 );
 
 CREATE TABLE qrtz_triggers
@@ -32,7 +38,7 @@ CREATE TABLE qrtz_triggers
     sched_name TEXT NOT NULL,
     trigger_name TEXT NOT NULL,
     trigger_group TEXT NOT NULL,
-    job_name  TEXT NOT NULL,
+    job_name TEXT NOT NULL,
     job_group TEXT NOT NULL,
     description TEXT NULL,
     next_fire_time BIGINT NULL,
@@ -44,10 +50,14 @@ CREATE TABLE qrtz_triggers
     end_time BIGINT NULL,
     calendar_name TEXT NULL,
     misfire_instr SMALLINT NULL,
+    misfire_orig_fire_time BIGINT NULL,
+    execution_group VARCHAR(200) NULL,
+    preferred_node VARCHAR(200) NULL,
+    preferred_node_auto BOOL NOT NULL DEFAULT FALSE,
     job_data BYTEA NULL,
-    PRIMARY KEY (sched_name,trigger_name,trigger_group),
-    FOREIGN KEY (sched_name,job_name,job_group)
-        REFERENCES qrtz_job_details(sched_name,job_name,job_group)
+    PRIMARY KEY (sched_name, trigger_name, trigger_group),
+    FOREIGN KEY (sched_name, job_name, job_group)
+        REFERENCES qrtz_job_details (sched_name, job_name, job_group)
 );
 
 CREATE TABLE qrtz_simple_triggers
@@ -58,16 +68,17 @@ CREATE TABLE qrtz_simple_triggers
     repeat_count BIGINT NOT NULL,
     repeat_interval BIGINT NOT NULL,
     times_triggered BIGINT NOT NULL,
-    PRIMARY KEY (sched_name,trigger_name,trigger_group),
-    FOREIGN KEY (sched_name,trigger_name,trigger_group)
-        REFERENCES qrtz_triggers(sched_name,trigger_name,trigger_group) ON DELETE CASCADE
+    PRIMARY KEY (sched_name, trigger_name, trigger_group),
+    FOREIGN KEY (sched_name, trigger_name, trigger_group)
+        REFERENCES qrtz_triggers (sched_name, trigger_name, trigger_group)
+        ON DELETE CASCADE
 );
 
-CREATE TABLE QRTZ_SIMPROP_TRIGGERS
+CREATE TABLE qrtz_simprop_triggers
 (
     sched_name TEXT NOT NULL,
-    trigger_name TEXT NOT NULL ,
-    trigger_group TEXT NOT NULL ,
+    trigger_name TEXT NOT NULL,
+    trigger_group TEXT NOT NULL,
     str_prop_1 TEXT NULL,
     str_prop_2 TEXT NULL,
     str_prop_3 TEXT NULL,
@@ -80,9 +91,10 @@ CREATE TABLE QRTZ_SIMPROP_TRIGGERS
     bool_prop_1 BOOL NULL,
     bool_prop_2 BOOL NULL,
     time_zone_id TEXT NULL,
-    PRIMARY KEY (sched_name,trigger_name,trigger_group),
-    FOREIGN KEY (sched_name,trigger_name,trigger_group)
-        REFERENCES qrtz_triggers(sched_name,trigger_name,trigger_group) ON DELETE CASCADE
+    PRIMARY KEY (sched_name, trigger_name, trigger_group),
+    FOREIGN KEY (sched_name, trigger_name, trigger_group)
+        REFERENCES qrtz_triggers (sched_name, trigger_name, trigger_group)
+        ON DELETE CASCADE
 );
 
 CREATE TABLE qrtz_cron_triggers
@@ -92,9 +104,10 @@ CREATE TABLE qrtz_cron_triggers
     trigger_group TEXT NOT NULL,
     cron_expression TEXT NOT NULL,
     time_zone_id TEXT,
-    PRIMARY KEY (sched_name,trigger_name,trigger_group),
-    FOREIGN KEY (sched_name,trigger_name,trigger_group)
-        REFERENCES qrtz_triggers(sched_name,trigger_name,trigger_group) ON DELETE CASCADE
+    PRIMARY KEY (sched_name, trigger_name, trigger_group),
+    FOREIGN KEY (sched_name, trigger_name, trigger_group)
+        REFERENCES qrtz_triggers (sched_name, trigger_name, trigger_group)
+        ON DELETE CASCADE
 );
 
 CREATE TABLE qrtz_blob_triggers
@@ -103,24 +116,25 @@ CREATE TABLE qrtz_blob_triggers
     trigger_name TEXT NOT NULL,
     trigger_group TEXT NOT NULL,
     blob_data BYTEA NULL,
-    PRIMARY KEY (sched_name,trigger_name,trigger_group),
-    FOREIGN KEY (sched_name,trigger_name,trigger_group)
-        REFERENCES qrtz_triggers(sched_name,trigger_name,trigger_group) ON DELETE CASCADE
+    PRIMARY KEY (sched_name, trigger_name, trigger_group),
+    FOREIGN KEY (sched_name, trigger_name, trigger_group)
+        REFERENCES qrtz_triggers (sched_name, trigger_name, trigger_group)
+        ON DELETE CASCADE
 );
 
 CREATE TABLE qrtz_calendars
 (
     sched_name TEXT NOT NULL,
-    calendar_name  TEXT NOT NULL,
+    calendar_name TEXT NOT NULL,
     calendar BYTEA NOT NULL,
-    PRIMARY KEY (sched_name,calendar_name)
+    PRIMARY KEY (sched_name, calendar_name)
 );
 
 CREATE TABLE qrtz_paused_trigger_grps
 (
     sched_name TEXT NOT NULL,
     trigger_group TEXT NOT NULL,
-    PRIMARY KEY (sched_name,trigger_group)
+    PRIMARY KEY (sched_name, trigger_group)
 );
 
 CREATE TABLE qrtz_fired_triggers
@@ -138,7 +152,8 @@ CREATE TABLE qrtz_fired_triggers
     job_group TEXT NULL,
     is_nonconcurrent BOOL NOT NULL,
     requests_recovery BOOL NULL,
-    PRIMARY KEY (sched_name,entry_id)
+    execution_group VARCHAR(200) NULL,
+    PRIMARY KEY (sched_name, entry_id)
 );
 
 CREATE TABLE qrtz_scheduler_state
@@ -147,24 +162,21 @@ CREATE TABLE qrtz_scheduler_state
     instance_name TEXT NOT NULL,
     last_checkin_time BIGINT NOT NULL,
     checkin_interval BIGINT NOT NULL,
-    PRIMARY KEY (sched_name,instance_name)
+    PRIMARY KEY (sched_name, instance_name)
 );
 
 CREATE TABLE qrtz_locks
 (
     sched_name TEXT NOT NULL,
-    lock_name  TEXT NOT NULL,
-    PRIMARY KEY (sched_name,lock_name)
+    lock_name TEXT NOT NULL,
+    PRIMARY KEY (sched_name, lock_name)
 );
 
-create index idx_qrtz_j_req_recovery on qrtz_job_details(requests_recovery);
-create index idx_qrtz_t_next_fire_time on qrtz_triggers(next_fire_time);
-create index idx_qrtz_t_state on qrtz_triggers(trigger_state);
-create index idx_qrtz_t_nft_st on qrtz_triggers(next_fire_time,trigger_state);
-create index idx_qrtz_ft_trig_name on qrtz_fired_triggers(trigger_name);
-create index idx_qrtz_ft_trig_group on qrtz_fired_triggers(trigger_group);
-create index idx_qrtz_ft_trig_nm_gp on qrtz_fired_triggers(sched_name,trigger_name,trigger_group);
-create index idx_qrtz_ft_trig_inst_name on qrtz_fired_triggers(instance_name);
-create index idx_qrtz_ft_job_name on qrtz_fired_triggers(job_name);
-create index idx_qrtz_ft_job_group on qrtz_fired_triggers(job_group);
-create index idx_qrtz_ft_job_req_recovery on qrtz_fired_triggers(requests_recovery);
+CREATE INDEX idx_qrtz_j_g_n ON qrtz_job_details (sched_name, job_group, job_name);
+CREATE INDEX idx_qrtz_t_j ON qrtz_triggers (sched_name, job_name, job_group);
+CREATE INDEX idx_qrtz_t_c ON qrtz_triggers (sched_name, calendar_name);
+CREATE INDEX idx_qrtz_t_g_n ON qrtz_triggers (sched_name, trigger_group, trigger_name);
+CREATE INDEX idx_qrtz_t_nft_st ON qrtz_triggers (sched_name, trigger_state, next_fire_time);
+CREATE INDEX idx_qrtz_ft_inst_job_req_rcvry ON qrtz_fired_triggers (sched_name, instance_name, requests_recovery);
+CREATE INDEX idx_qrtz_ft_j_g ON qrtz_fired_triggers (sched_name, job_name, job_group);
+CREATE INDEX idx_qrtz_ft_t_g ON qrtz_fired_triggers (sched_name, trigger_name, trigger_group);
